+++ src/network/apis/answerApi.ts
@@ -1,4 +1,4 @@
-import { publicInstance } from './config';
+import { publicInstance } from '../config';
 
 export const postAnswerApi = async ({
   postIdx,
